feat: init memory & add cpu/vm struct

+++ src/arch/aarch64/platform.rs
@@ -1,7 +1,7 @@
 pub struct GicDesc {
-    pub gicd_addr: u64,
-    pub gicc_addr: u64,
-    pub gich_addr: u64,
-    pub gicv_addr: u64,
-    pub maintenance_int_id: u64,
+    pub gicd_addr: usize,
+    pub gicc_addr: usize,
+    pub gich_addr: usize,
+    pub gicv_addr: usize,
+    pub maintenance_int_id: usize,
 }
